rename file by ValidationOfTheRegistrationForm and package.lock.json erased
