+++ src/middlewares/validations/ValidationOfTheRegistrationForm.js
