+++ src/routes/users.js
@@ -1,12 +1,12 @@
 const express = require("express");
 const router = express.Router();
 const userController = require('../controllers/users-controllers');
-const validationFromRegister = require("../middlewares/validations/validationFromRegister");
+const ValidationOfTheRegistrationForm = require("../middlewares/validations/ValidationOfTheRegistrationForm");
 
 
 router.get("/users");
 router.post("/login", userController.login)
-router.post("/register", validationFromRegister, userController.register)
+router.post("/register", ValidationOfTheRegistrationForm, userController.register)
